Updates for full UC Run
Clip_CDL Modified Clip_CDL to bring along colormap file.

Edited Et_Demands_Zonal_Stats to accept more than 255 zones when using
ZonalHistorgram function. Now uses loop to steps through zones in 250
chunks.

+++ et-demands/prep/clip_cdl_raster.py
@@ -105,7 +105,10 @@ def main(gis_ws, cdl_input_ws, cdl_year='', overwrite_flag=False,
                 ['-projwin'] + str(output_ullr).split() +
                 ['-a_ullr'] + str(output_ullr).split() +
                 [cdl_input_path, cdl_output_path])
+            if os.path.isfile(cdl_input_path.replace('.img', '.img.vat.dbf')):
                 shutil.copyfile(
+                    cdl_input_path.replace('.img', '.img.vat.dbf'),
+                    cdl_output_path.replace('.img', '.img.vat.dbf')
                 )
             # , '-a_srs', 'output_proj'
             # subprocess.call(
